+++ index.js
@@ -1,84 +1,80 @@
 const path = require('path');
 const express = require('express');
-const request = require('request');
 const serveStatic = require('serve-static');
 const cache = require('memory-cache');
+const nconf = require('nconf').argv().env().file({ file: __dirname + '/config.json' });
 
 const { getDependencies } = require('./lib/starbuck');
 const { getBadge } = require('./lib/util');
+const github = require('./lib/github')(nconf.get('github'));
+const gitlab = require('./lib/gitlab')(nconf.get('gitlab'));
+
+const supportedServices = ['gitlab', 'github'];
+const supportedBadges = ['dev-status', 'status', 'peer-status'];
 
 const app = express();
 const port = process.env.PORT || 8000;
+const asyncMiddleware = (fn) => {
+	return (req, res, next) => {
+		Promise.resolve(fn(req, res, next)).catch(next);
+	};
+};
 
 app.use(serveStatic('dist'));
 
-app.get('/api/github/:owner/:repo', (req, res) => {
-	const { owner, repo } = req.params;
-	if (cache.get(`github:${owner}:${repo}`)) {
-		return res.status(200).send(cache.get(`github:${owner}:${repo}`));
+app.get('/api/:service/:owner/repos', asyncMiddleware(async (req, res) => {
+	const { service, owner } = req.params;
+	if (cache.get(`${service}:${owner}:repos`)) {
+		return res.status(200).send(JSON.stringify(cache.get(`${service}:${owner}:repos`)));
 	}
 
-	request({
-		method: 'GET',
-		json: true,
-		url: `https://api.github.com/repos/${owner}/${repo}/contents/package.json`,
-		headers: {
-			'User-Agent': 'request'
-		}
-	}, (err, results) => {
-		const pack = JSON.parse(Buffer.from(results.body.content, 'base64').toString('utf8'));
-
-		getDependencies(pack, {
-			npm: 'http://registry.npmjs.org'
-		}, (error, result) => {
-			if (error) return res.status(500).send(error);
-			let response = JSON.stringify(Object.assign({
-				starbuck: result
-			}, pack));
-
-			cache.put(`github:${owner}:${repo}`, response, 900000);
-
-			res.status(200).send(response);
-		});
-	});
-});
-
-app.get('/api/gitlab/:owner/:repo', (req, res) => {
-	const { owner, repo } = req.params;
-	if (cache.get(`gitlab:${owner}:${repo}`)) {
-		return res.status(200).send(cache.get(`gitlab:${owner}:${repo}`));
+	let repos = {};
+	if(service === 'github') {
+		repos = await github.getRepos(owner);
+	} else {
+		repos = await gitlab.getRepos(owner);
 	}
-	const id = encodeURIComponent(`${owner}/${repo}`);
 
-	request({
-		method: 'GET',
-		json: true,
-		url: `https://www.gitlab.com/api/v3/projects/${id}/repository/files/package.json?ref=master`,
-		headers: {
-			'User-Agent': 'request'
-		}
-	}, (err, results) => {
-		const pack = JSON.parse(Buffer.from(results.body.content, 'base64').toString('utf8'));
+	cache.put(`${service}:${owner}:repos`, repos, 900000);
 
-		getDependencies(pack, {
-			npm: 'http://registry.npmjs.org'
-		}, (error, result) => {
-			if (error) return res.status(500).send(error);
-			let response = JSON.stringify(Object.assign({
-				starbuck: result
-			}, pack));
+	res.send(repos);
+}));
 
-			cache.put(`gitlab:${owner}:${repo}`, response, 900000);
+app.get('/api/:service/:owner/:repo', asyncMiddleware(async (req, res) => {
+	const { service, owner, repo } = req.params;
+	if (supportedServices.indexOf(service) === -1) {
+		return res.status(500).send({ error: 'service not supported' });
+	}
 
-			res.status(200).send(response);
-		});
+	if (cache.get(`${service}:${owner}:${repo}`)) {
+		return res.status(200).send(JSON.stringify(cache.get(`${service}:${owner}:${repo}`)));
+	}
+
+	let pack = {};
+	if(service === 'github') {
+		pack = await github.getPackage(owner, repo);
+	} else {
+		pack = await gitlab.getPackage(owner, repo);
+	}
+
+	const dependencies = await getDependencies(pack, {
+		npm: 'http://registry.npmjs.org'
 	});
-});
+	const response = Object.assign({ starbuck: dependencies }, pack);
 
-app.get('/badge/gitlab/:owner/:repo/:status.svg', (req, res) => {
-	const allowed = ['dev-status', 'status', 'peer-status'];
-	const { owner, repo, status } = req.params;
-	if (allowed.indexOf(status) === -1) {
+	cache.put(`${service}:${owner}:${repo}`, response, 900000);
+
+	res.status(200).send(JSON.stringify(response));
+}));
+
+app.get('/badge/:service/:owner/:repo/:status.svg', asyncMiddleware(async (req, res) => {
+	const { service, owner, repo, status } = req.params;
+
+	if (supportedServices.indexOf(service) === -1) {
+		return res.status(500).send({ error: 'service not supported' });
+	}
+
+	if (supportedBadges.indexOf(status) === -1) {
 		return res.sendFile(getBadge('unknown', ''), {
 			lastModified: false,
 			etag: false,
@@ -89,140 +85,61 @@ app.get('/badge/gitlab/:owner/:repo/:status.svg', (req, res) => {
 		});
 	}
 
-	function get(callback) {
-		if (cache.get(`gitlab:${owner}:${repo}`)) {
-			return callback(JSON.parse(cache.get(`gitlab:${owner}:${repo}`)));
+	async function get() {
+		if (cache.get(`${service}:${owner}:${repo}`)) {
+			return cache.get(`${service}:${owner}:${repo}`);
 		}
-		const id = encodeURIComponent(`${owner}/${repo}`);
 
-		request({
-			method: 'GET',
-			json: true,
-			url: `https://www.gitlab.com/api/v3/projects/${id}/repository/files/package.json?ref=master`,
-			headers: {
-				'User-Agent': 'request',
-			}
-		}, (err, results) => {
-			const pack = JSON.parse(Buffer.from(results.body.content, 'base64').toString('utf8'));
-
-			getDependencies(pack, {
-				npm: 'http://registry.npmjs.org'
-			}, (error, result) => {
-				if (error) return res.status(500).send(error);
-				let response = JSON.stringify(Object.assign({
-					starbuck: result
-				}, pack));
-
-				cache.put(`gitlab:${owner}:${repo}`, response, 900000);
-				callback(JSON.parse(response));
-			});
+		let pack = {};
+		if(service === 'github') {
+			pack = await github.getPackage(owner, repo);
+		} else {
+			pack = await gitlab.getPackage(owner, repo);
+		}
+		const dependencies = await getDependencies(pack, {
+			npm: 'http://registry.npmjs.org'
 		});
+		const response = Object.assign({ starbuck: dependencies }, pack);
+
+		cache.put(`${service}:${owner}:${repo}`, response, 900000);
+
+		return response;
 	}
 
-	get((response) => {
-		let dep = {
-			'dev-status': 'devDependencies',
-			'status': 'dependencies',
-			'peer-status': 'peerDependencies'
-		}[status];
+	const response = await get();
+	let dep = {
+		'dev-status': 'devDependencies',
+		'status': 'dependencies',
+		'peer-status': 'peerDependencies'
+	}[status];
 
-		let dependencies = response.starbuck[dep];
+	let dependencies = response.starbuck[dep];
 
-		let type;
-		if (Object.keys(dependencies).length === 0) {
-			type = 'none';
-		} else if (Object.keys(dependencies).filter((d) => dependencies[d].needsUpdating).length > 0) {
-			type = 'notsouptodate';
-		} else {
-			type = 'uptodate';
+	let type;
+	if (Object.keys(dependencies).length === 0) {
+		type = 'none';
+	} else if (Object.keys(dependencies).filter((d) => dependencies[d].needsUpdating).length > 0) {
+		type = 'notsouptodate';
+	} else {
+		type = 'uptodate';
+	}
+
+	res.sendFile(getBadge(type, dep), {
+		lastModified: false,
+		etag: false,
+		headers: {
+			'Cache-Control': 'no-cache, no-store, must-revalidate',
+			'Expires': new Date().toUTCString()
 		}
-
-		res.sendFile(getBadge(type, dep), {
-			lastModified: false,
-			etag: false,
-			headers: {
-				'Cache-Control': 'no-cache, no-store, must-revalidate',
-				'Expires': new Date().toUTCString()
-			}
-		});
 	});
-});
-
-app.get('/badge/github/:owner/:repo/:status.svg', (req, res) => {
-	const allowed = ['dev-status', 'status', 'peer-status'];
-	const { owner, repo, status } = req.params;
-	if (allowed.indexOf(status) === -1) {
-		return res.sendFile(getBadge('unknown', ''), {
-			lastModified: false,
-			etag: false,
-			headers: {
-				'Cache-Control': 'no-cache, no-store, must-revalidate',
-				'Expires': new Date().toUTCString()
-			}
-		});
-	}
-
-	function get(callback) {
-		if (cache.get(`github:${owner}:${repo}`)) {
-			callback(JSON.parse(cache.get(`github:${owner}:${repo}`)));
-		} else {
-			request({
-				method: 'GET',
-				json: true,
-				url: `https://api.github.com/repos/${owner}/${repo}/contents/package.json`,
-				headers: {
-					'User-Agent': 'request'
-				}
-			}, (err, results) => {
-				const pack = JSON.parse(Buffer.from(results.body.content, 'base64').toString('utf8'));
-
-				getDependencies(pack, {
-					npm: 'http://registry.npmjs.org'
-				}, (error, result) => {
-					if (error) return res.status(500).send(error);
-					let response = JSON.stringify(Object.assign({
-						starbuck: result
-					}, pack));
-
-					cache.put(`github:${owner}:${repo}`, response, 900000);
-
-					callback(JSON.parse(response));
-				});
-			});
-		}
-	}
-
-	get((response) => {
-		let dep = {
-			'dev-status': 'devDependencies',
-			'status': 'dependencies',
-			'peer-status': 'peerDependencies'
-		}[status];
-
-		let dependencies = response.starbuck[dep];
-
-		let type;
-		if (Object.keys(dependencies).length === 0) {
-			type = 'none';
-		} else if (Object.keys(dependencies).filter((d) => dependencies[d].needsUpdating).length > 0) {
-			type = 'notsouptodate';
-		} else {
-			type = 'uptodate';
-		}
-
-		res.sendFile(getBadge(type, dep), {
-			lastModified: false,
-			etag: false,
-			headers: {
-				'Cache-Control': 'no-cache, no-store, must-revalidate',
-				'Expires': new Date().toUTCString()
-			}
-		});
-	});
-});
+}));
 
 app.get('*', (req, res) => {
 	res.sendFile(path.resolve(__dirname, 'dist', 'index.html'));
+});
+
+app.use(function (err, req, res) {
+	res.status(500).send({ error: err.toString() });
 });
 
 app.listen(port, () => {