+++ stories/index.js
@@ -2,3 +2,4 @@ import 'psychic-ui/dist/psychic-min.css';
 import '../src/style.css';
 
 import './dependenciesTable';
+import './user';