+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.0 (10/16/2017)
+
+- updates dependencies
+- fixes case sensitive bug in `src/router.js`
+- add page to show all users repos `/:service/:owner`
+- cleans up router logic
+- adds the ability to select the table using hashes (github/gabrielcsapo/compress-object/peerDependencies)
+- properly handles errors on the server and client side
+- make github and gitlab endpoints configurable (be able to add authorization credentials)
+
 # 0.0.0 (10/09/2017)
 
 - adds storybook and basic doc page